Prevent revisions to be persisted in localStorage
Created a new `state/revisions` folder containing only actions and
reducers related to `revisions`, in order to separate them from
`settings` and to not persist them in `localStorage`.
