+++ lib/state/revisions/actions.js
@@ -1,0 +1,3 @@
+export const setRevisionState = revisionState => ( {
+	type: 'setRevisionState', revisionState,
+} );
