+++ lib/state/settings/actions.js
@@ -49,7 +49,3 @@ export const setAccountName = accountName => ( {
 export const toggleLayoutType = layoutType => ( {
 	type: 'toggleLayoutType', layoutType,
 } );
-
-export const setRevisionState = revisionState => ( {
-	type: 'setRevisionState', revisionState,
-} );
