+++ lib/state/settings/reducer.js
@@ -65,17 +65,6 @@ const layoutType = ( state = '3-col', action ) => {
 	return action.layoutType;
 };
 
-const revisionState = ( state = { revision: null, isViewingRevisions: false }, action ) => {
-	if ( 'setRevisionState' !== action.type ) {
-		return state;
-	}
-	console.log( action );
-	return {
-		...state,
-		...action.revisionState,
-	};
-};
-
 export default combineReducers( {
 	accountName,
 	fontSize,
@@ -85,5 +74,4 @@ export default combineReducers( {
 	sortReversed,
 	theme,
 	layoutType,
-	revisionState,
 } );
